typo - removing an extraneous space
The space is unfortunately in a command line which doesn't
expect a space just there (-- httpport option to rpm) and therefore
breaks. This was already fixed upstream, but we're quite a bit
adrift from upstream :(. In order to get the simplest possible fix
for now, I'm just applying this simple fix rather than working
on merging the two branches.